requirements.txt
appdirs==1.4.4
backcall==0.2.0
black==21.6b0
click==8.0.1
cycler==0.10.0
debugpy==1.3.0
decorator==4.4.2
imageio==2.9.0
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.9.2
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mypy==0.910
mypy-extensions==0.4.3
networkx==2.5.1
numpy==1.21.0
opencv-python==4.5.2.54
OpenEXR==1.3.2
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pkg-resources==0.0.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycairo==1.20.1
pydantic==1.8.2
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
pyzmq==22.1.0
regex==2021.7.6
scikit-image==0.18.2
scipy==1.7.0
six==1.16.0
tifffile==2021.7.2
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typing-extensions==3.10.0.0
wcwidth==0.2.5
